Renamed package, fixed issues with overriding bindViewHolder funs
