+++ WooCommerce/src/main/kotlin/com/woocommerce/android/widgets/sectionedrecyclerview/SectionedRecyclerViewAdapter.kt
@@ -1,11 +1,11 @@
-package com.woocommerce.android.widgets
+package com.woocommerce.android.widgets.sectionedrecyclerview
 
 import android.support.annotation.VisibleForTesting
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.woocommerce.android.widgets.Section.State
+import com.woocommerce.android.widgets.sectionedrecyclerview.Section.State
 import java.util.HashMap
 import java.util.LinkedHashMap
 import java.util.UUID
